split: add count of replicas actively sampling per store
One of our goals with adding more hotspot telemetry is to better
understand what's happening to a cluster when it has a hotspot. Today this is
possible real time, but information is limited when trying to understand
hotspots from the past.

We currently have a log for the hot ranges in a cluster, which can be enabled
to periodically report the hot ranges, but to limit the output it runs
infrequently, and therefore is likely to miss temporal, or short lived
hotspots.

In the replica deciders, there already exists some functionality for
determining when a specific replica is the target of an unbalanced
portion of the system's load. What this change does is allow for other
parts of the system (namely the hot range logger) to subscribe to when
that tipping point is reached.

The following change will link the hot range logger to this new
notification system, so that temporal hotspots can be better examined.

Fixes: #138672
Epic: CRDB-43150

Release note: none