structlogging: fix flaky hot range job test
A change introduced in 80d454f fixed the exit behavior of the hot range
logging job so that it would restart if it were an app tenant, and exit
gracefully if it were a system tenant.

However, due to the stochastic nature of the test deployment, when there
is only a system tenant, it would mistakenly believe
`ApplicationLayer().Codec` would not be the system codec, and therefore
fail the test.

This PR fixes this test, by affixing the deployment mode of the test to
include an app tenant.

Fixes: #149977
Epic: none
Release note: none