sql,sqlstats: Support fingerprint normalization
First commit to start work for fingerprint normalization.
This commit adds functions to the SQLStatsController for
creating statement and transaction fingerprints. This
interface exists on the eval context so it should be
accessible in most, if not all, of the sql layer.

The new methods, CreateStatementFingerprint and
CreateTransactionFingerprint should be used to create
fingerprints because it will create the ids, cache them,
and store them in a backing system table

Epic: None
Release note: None